Conflicts:
	workspace/Assembly-CSharp-vs.csproj
	workspace/Assembly-CSharp.csproj
	workspace/Assets/Scripts/data/obj.meta
	workspace/Library/CurrentLayout.dwlt
	workspace/Library/assetDatabase3
	workspace/Library/expandedItems
	workspace/Library/guidmapper
	workspace/Library/metadata/00/00614899600137c41a9fd2764092a5a3
	workspace/Library/metadata/05/05ee0b17724aa5c4d971c0171755c9ea
	workspace/Library/metadata/08/084f0033aa719e645baefc8b4b3c04c3
	workspace/Library/metadata/1f/1fa5d394e0e2aa441b7c794fa2e1326b
	workspace/Library/metadata/22/22885f8fa14c55d4a9dd119d2daa6134
	workspace/Library/metadata/2f/2f1c2d9a4d8fedc41a535f6b561b3dba
	workspace/Library/metadata/2f/2f9027f720d80f440b2eb8a2c1ec2f64
	workspace/Library/metadata/3f/3f631d4394c477f439e2adfa45870b41
	workspace/Library/metadata/44/44f4e134d5ddd664c977a8888b091ab1
	workspace/Library/metadata/5e/5e29e2b71e2adee4ea8dbae85e2b0665
	workspace/Library/metadata/80/803e33c72f3c2cf45bd2b0f06c037c5f
	workspace/Library/metadata/86/864e2eb40caa51e418a2e9b20da3b9ac
	workspace/Library/metadata/b7/b7ddefdd959352e49824a8128fd400e4
	workspace/Library/metadata/bb/bb49eecbd728f1543a6ab49762719d13
	workspace/Library/metadata/cd/cd12e2eb7fc4995489bce4251c679b7c
	workspace/Library/metadata/ea/ea4070d7c6a8a1b48a241463504f3abc
	workspace/Library/metadata/fe/fe8b8000e48de3143bbe33da7015c28c
	workspace/Library/shadercompiler-1.log
	workspace/workspace-csharp.sln
	workspace/workspace.sln
	workspace/workspace.userprefs

+++ workspace/Assets/Scripts/Game/data/monsterDataLoad.cs
@@ -39,7 +39,7 @@ public class monsterDataLoad : CSVParser {
 		tempData.magicAtk 		    = Convert.ToInt32( inputData[count++] );
 		tempData.moveSpd 		    = (float)Convert.ToDouble( inputData [count++] ); 
 		tempData.atkSpd 			= (float)Convert.ToDouble( inputData [count++] );
-		tempData.shotSize 		    = Convert.ToInt32( inputData[count++] );
+		tempData.shotSize 		    = (float)Convert.ToDouble( inputData [count++] );
 		tempData.shotSpd 		    = (float)Convert.ToDouble (inputData [count++] );
 		tempData.flightPattern 		= Convert.ToInt32( inputData[count++] );
 		tempData.gold 				= Convert.ToInt32( inputData[count++] );
